feat: Implement player feature

+++ core-utils/src/main/java/com/example/core_utils/util/PlayerUtil.kt
@@ -1,7 +1,7 @@
 package com.example.core_utils.util
 
 object PlayerUtil {
-    fun durationToString(duration: Int?): String {
+    fun durationToString(duration: Long?): String {
         if(duration == null) return "0:00"
         val minute = (duration/60).toString()
         var second = (duration % 60).toString()
